requirements.txt
numpy
pandas
scikit-learn
matplotlib
librosa
soundfile
torch
umap-learn
pytest
